Add karatsuba mutliplication algorithm
Adds an initial sketch of the algorithm. This version outlines a proof
of concept by using a set of small integer values to do the
multiplication. Really we would never do it like this because the
standard multiplcation algorithm is much more efficient when dealing
with these sorts of numbers.