Dump routine wip
Trying to untangle an knotty terneray expression from the original:

 OK, so there are 6 sets of strings, empty:"  ", branch:"|-", verti:"| ",
last:"`-", single: "---" and first: "-+-"

It happens in dump_tree() which is called recursively for each sub-tree.

There's an awful stacked ternary conditional operator
           out_string(lvl ==
                         level -
                         1 ? last ? sym->last_2 : sym->branch_2 : more[lvl +
                                                                       1] ?
                         sym->vert_2 : sym->empty_2);

if level == level-1 {
 if last {
    print `- # last
  } else {
    print |- # branch
  } else {
  if more[lvl+1] {
    print |  # vert
} else {
  print empty
}

The printing first/branch/last  is just checking to see if there are more
in the same-sub-tree or not e.g.
parent -+- first_proc(1)
            |- branch_proc(2)
            `- last_proc(3)

So that prints:
  first for proc1 (I am the first at this level but there are more)
  branch for proc2 (I am not the first and there are more)
  last for proc3 (I am not the first, but there are no more)