Initial implementation of anagram
This introduces a couple of simple datatypes and gets as far as
coding up a map to hold the key value pairs of words. Where key
is the signature that we will use to associate the words and
value is a linked list containing all words associated with that key.