Fix: two subtle bugs
1. If a map is used to store the count of the runes for building
the key in "initToken", then the order of the keys is not guaranteed.
This means that the structure of the key will be non-deterministic
(i.e. prone to random variance over time).
The function already sorts the runes in the word it receives as input.
If this sorted list of runes is deduplicated with "slices.Compact",
then it is possible to safely iterate over the list of runes, using
the items as keys for value lookups in the map.
2. If 'AnagramMap."Load"' receives a reader with lines containing multiple,
whitespace seperated words, it would still try to erroneosly process the line.
Really, we only want to process lines if they are actual _words_ from a dictionry.
There should be one per line. In this case, return an error upon finding
a line with multiple words in the input stream.